requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
greenlet==3.2.3
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
